+++ main.tf
@@ -19,6 +19,7 @@
 locals {
   create_policy = var.policy == null && length(var.policy_statements) > 0
   policy        = var.module_enabled ? (local.create_policy ? data.aws_iam_policy_document.policy[0].json : var.policy) : null
+  tags          = merge(var.module_tags, var.tags)
 }
 
 resource "aws_iam_policy" "policy" {
@@ -31,6 +32,7 @@ resource "aws_iam_policy" "policy" {
 
   policy = local.policy
 
+  tags       = local.tags
   depends_on = [var.module_depends_on]
 }